pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waybackproxy"
version = "2.0.0"
description = "Retro-friendly HTTP proxy for Internet Archive Wayback Machine"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "izzy-gm"}
]
dependencies = [
    "urllib3>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
# Hardware UI dependencies (Raspberry Pi only)
hardware = [
    "rpi-lgpio>=0.4.0",
    "smbus2>=0.4.2",
]

[project.scripts]
waybackproxy = "waybackproxy.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
